server/requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
crispy-tailwind==0.5.0
Django==4.1.6
django-crispy-forms==1.14.0
django-turbo-response==0.0.52
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
python-dateutil==2.8.2
python-slugify==8.0.0
python-webpack-boilerplate==1.0.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
urllib3==1.26.14
